--- my-react-app/src/TodoList/Index.js
@@ -1,14 +1,0 @@
-import react from "react";
-import './TodoList.css'
-
-function TodoList(props){
-    return(
-        <section>
-            <ul>
-                {props.children}
-            </ul>
-        </section>
-    );
-}
-
-export { TodoList};
