+++ src/TodoCounter/Index.js
@@ -1,4 +1,3 @@
-import React from "react";
 import './TodoCounter.css';
 
 function TodoCounter({total,completed}) {
