+++ src/TodoItem/Index.js
@@ -1,0 +1,22 @@
+import "./TodoItem.css";
+
+function TodoItem(props) {
+  return (
+    <li className="TodoItem">
+      <span
+        className={`Icon Icon-check ${props.completed && "Icon-check--active"}`}
+        onClick={props.onComplete}
+      >
+        √
+      </span>
+      <p className={`TodoItem-p ${props.completed && "TodoItem-p--complete"}`}>
+        {props.text}
+      </p>
+      <span className="Icon Icon-delete" onClick={props.onDelete}>
+        X
+      </span>
+    </li>
+  );
+}
+
+export { TodoItem };
