+++ src/TodoList/Index.js
@@ -1,0 +1,8 @@
+import "./TodoList.css";
+
+// No hay necesidad de usar un section, las etiquitas UL funcionan como contenedor
+function TodoList(props) {
+  return <ul>{props.children}</ul>;
+}
+
+export { TodoList };
